Make 'batchSize' configurable, add a gauge.
In order to determine the effect of changing the batch size for the HBase, we
need to be able to modify it, ideally without re-compiling each time.  This
commit adds the ability to specify the batch size as a run-time parameter.

We also add a gauge so that we can check the current value of batch size at
run time, giving us a denominator to make sense of the flush times.