Add a "replay" consumer
This consumer reads from a kafka queue, re-creates the original
requests, and submits them as new HTTP requests.

It is intended to help in testing the stage environment using
production data.

This inital version supports optional sampling of the queued records and
the ability to generate new keys for each request or copy the source keys.